// src/types/movie.ts
export type IMovie = {
    Title: string;
    Year: string;
    imdbID: string;
    Type: string;
    Poster: string;
    Loading?: boolean;
}

export type IMovieResponse = {
    Search: IMovie[];
    totalResults: string;
    Response: string;
}

export type IMovieDetails = {
    "Title": string;
    "Year": string;
    "Rated": string;
    "Released": string;
    "Runtime": string;
    "Genre": string;
    "Director": string;
    "Writer": string;
    "Actors": string;
    "Plot": string;
    "Language": string;
    "Country": string;
    "Awards": string;
    "Poster": string;
    "Ratings": { "Source": string; "Value": string; }[],
    "Metascore": string;
    "imdbRating": string;
    "imdbVotes": string;
    "imdbID": string;
    "Type": string;
    "DVD": string;
    "BoxOffice": string;
    "Production": string;
    "Website": string;
    "Response": string
}